requirements.txt
Flask==3.0.0
requests==2.31.0
reportlab==4.0.7
python-docx==1.1.0
supabase==2.28.0
websockets==15.0.1
python-dotenv==1.0.1
